pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymage"
version = "0.0.2"
authors = [
  { name="Fabrice Gaillard", email="fgaillard@w3architect.com" },
]
description = "polymage : a multimodal agent python library"
readme = "README.md"
license = {text = "Apache 2.0"}
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.1",
    "inspector>=0.5.1.1",
    "openai>=2.11.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "setuptools>=80.9.0",
    "twine>=6.2.0",
]
